Rename app to zerocom
