+++ .flake8
@@ -1,7 +1,7 @@
 [flake8]
 max_line_length=120
 extend-ignore=E203 # flake8 has conflicting handling with black here
-application-import-names=app,tests
+application-import-names=zerocom,tests
 exclude=.venv,.git,.cache
 ignore=
     ANN002, # *args
