+++ zerocom/protocol/utils.py
